+++ rpc/src/v1/impls/eth_filter.rs
@@ -21,7 +21,7 @@ use std::sync::{Arc, Weak, Mutex};
 use std::collections::HashSet;
 use jsonrpc_core::*;
 use util::numbers::*;
-use ethminer::{MinerService};
+use ethminer::MinerService;
 use ethcore::filter::Filter as EthcoreFilter;
 use ethcore::client::{BlockChainClient, BlockID};
 use v1::traits::EthFilter;
