+++ rpc/src/v1/impls/mod.rs
@@ -28,6 +28,7 @@ macro_rules! take_weak {
 mod web3;
 mod eth;
 mod eth_filter;
+mod eth_signing;
 mod net;
 mod personal;
 mod ethcore;
@@ -35,8 +36,9 @@ mod traces;
 mod rpc;
 
 pub use self::web3::Web3Client;
-pub use self::eth::{EthClient, EthSigningUnsafeClient};
+pub use self::eth::EthClient;
 pub use self::eth_filter::EthFilterClient;
+pub use self::eth_signing::EthSigningUnsafeClient;
 pub use self::net::NetClient;
 pub use self::personal::PersonalClient;
 pub use self::ethcore::EthcoreClient;
